# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: School Report Card Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subjects:
    get:
      operationId: listSubjects
      tags: [subjects]
      summary: List subjects
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
    post:
      operationId: createSubject
      tags: [subjects]
      summary: Create subject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
  /subjects/{id}:
    patch:
      operationId: updateSubject
      tags: [subjects]
      summary: Update subject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    delete:
      operationId: deleteSubject
      tags: [subjects]
      summary: Delete subject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  # ── USERS ─────────────────────────────────────────────────────────────────
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List students
      parameters:
        - name: classId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  # ── ASSESSMENT COMPONENTS ─────────────────────────────────────────────────
...
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
...
    Subject:
      type: object
      required: [id, name, code]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
    SubjectInput:
      type: object
      required: [name, code]
      properties:
        name:
          type: string
        code:
          type: string
    SubjectUpdate:
      type: object
      properties:
        name:
          type: string
        code:
          type: string

    # ── User ──────────────────────────────────────────────────────────────
...
    Student:
      type: object
      required: [id, studentIdNumber, fullName, classId]
      properties:
        id:
          type: integer
        studentIdNumber:
          type: string
        fullName:
          type: string
        dateOfBirth:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        classId:
          type: integer
        className:
          type: string
        guardianName:
          type: ["string", "null"]
        guardianPhone:
          type: ["string", "null"]
        admissionDate:
          type: ["string", "null"]
    StudentInput:
      type: object
      required: [studentIdNumber, fullName, classId]
      properties:
        studentIdNumber:
          type: string
        fullName:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
          enum: [male, female]
        classId:
          type: integer
        guardianName:
          type: string
        guardianPhone:
          type: string
        admissionDate:
          type: string
    StudentUpdate:
      type: object
      properties:
        fullName:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
        classId:
          type: integer
        guardianName:
          type: string
        guardianPhone:
          type: string

    # ── Assessment Components ─────────────────────────────────────────────
...